// grading-admin-frontend/apps/admin/src/lib/api/v1.d.ts
export interface components {
    schemas: {
// ...
        Assignment: {
            /** @example 999-9999-9999-99 */
            id?: string;
            /** @example Assessment 1 */
            name: string;
            /** @example Quiz */
            category: string;
            /**
             * Format: int64
             * @example 1245678
             */
            canvas_id: number;
            /**
             * Format: double
             * @example 15
             */
            points: number;
            /** @example Gradescope */
            external_service?: string;
            /** @example 14 */
            external_points?: number;
            /**
             * Format: date-time
             * @example 2020-01-15T12:00:00.000Z
             */
            due_date: string;
            /**
             * Format: date-time
             * @example 2020-01-01T12:00:00.000Z
             */
            unlock_date: string;
            /** @example true */
            enabled: boolean;
            /** @example false */
            group_assignment?: boolean;
            /** @example false */
            attention_required?: boolean;
            /** @example false */
            frozen?: boolean;
        };
// ...
        Course: {
// ...
            assignments?: components["schemas"]["Assignment"][];
            /** @example [
             *       "fall.2020.excl.101.section.a"
             *     ] */
            sections?: string[];
            late_request_config: components["schemas"]["CourseLateRequestConfig"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}